third_party/WebKit/LayoutTests/fast/dom/Window/getMatchedCSSRules-with-pseudo-elements-expected.txt

CONSOLE WARNING: line 20: document.getMatchedCSSRules() is deprecated and will be removed in M63, around December 2017. See https://www.chromestatus.com/features/4606972603138048 for more details.
Test for getMatchedCSSRules for Pseudo Elements

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS pseudoRules[0].selectorText is selectorText
PASS pseudoRules[0].style.cssText.indexOf("1px") != -1 is true
PASS successfullyParsed is true

TEST COMPLETE
